% Generated by roxygen2 (4.0.2): do not edit by hand
\name{a_find}
\alias{a_find}
\title{Find function for lambda expression}
\usage{
a_find(data, ..., right = FALSE, nomatch = NULL)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}
}
\value{
an index.
}
\description{
Find function for lambda expression
}

