% Generated by roxygen2 (4.0.2): do not edit by hand
\name{a_mapu}
\alias{a_mapu}
\title{Map function for lambda expression with unlisted output}
\usage{
a_mapu(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression.}
}
\value{
a vector.
}
\description{
Map function for lambda expression with unlisted output
}

