% Generated by roxygen2 (4.0.2): do not edit by hand
\name{a_reduce}
\alias{a_reduce}
\title{Reduce function for lambda expression}
\usage{
a_reduce(data, ..., init, right = FALSE, accumulate = FALSE)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(binary).}
}
\value{
a value.
}
\description{
Reduce function for lambda expression
}

