% Generated by roxygen2 (4.0.2): do not edit by hand
\name{a_takeWhile}
\alias{a_takeWhile}
\title{takeWhile function for lambda expression}
\usage{
a_takeWhile(data, ..., nomatch = NULL)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}
}
\value{
a vector.
}
\description{
takeWhile function for lambda expression
}
\examples{
library(lambdaR)
library(dplyr)
1:10 \%>\% a_takeWhile(x: x \%\% 2 == 0)
}

