\name{ApCat.sample}
\alias{ApCat.sample}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{
Sample Aperture Catalogue
}
\description{
A sample of the GAMA Aperture Catalouge, taken from the GAMA 12Hr Field.
This catalogue matches the sample SDSS image also supplied within the
LAMBDAR package sample data.
}
%- maybe also `usage' for other objects documented here.
\keyword{ Data }
