% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws-generic-params}
\alias{aws-generic-params}
\title{Shared params for AWS cli functions}
\arguments{
\item{aws_account_id}{Your 12-digit AWS account id}

\item{aws_region}{Your AWS region, e.g. \code{"ap-southeast-2"}}

\item{aws_ecr_repository_name}{Chosen name for your AWS ECR repository. It is recommended that
this be the same as your app name.}

\item{tag}{Optional image tag. If omited, "latest" will be used.}
}
\description{
Shared params for AWS cli functions
}
\keyword{internal}
