% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{build_dockerfile}
\alias{build_dockerfile}
\title{Create a Dockerfile}
\usage{
build_dockerfile()
}
\value{
Nothing is returned, but a Dockerfile will be written to disk.
}
\description{
Create a Dockerfile
}
