% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{build_image}
\alias{build_image}
\title{Build a container image from a config file}
\usage{
build_image()
}
\description{
Reads your \verb{_lambdar.yml} config file, creates a Dockerfile from it, and then attempts to create
a container from the Dockerfile using \verb{docker build}. You must have Docker installed for this to
work, otherwise it will throw an error.
}
