% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{build_yaml}
\alias{build_yaml}
\title{Build a `_lambdar.yml` file}
\usage{
build_yaml()
}
\description{
Scans your project folder for `.R` files, identified those containing `@lambda` tags, and writes
a `_lambdar.yml` config file containing relevant information.
}
