\name{createparfile}
\alias{createparfile}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{
Create Parameter File
}
\description{
Generate the parameter file for use in the MeasureFluxes procedure. All Parameter file
values can be sent to the procedure, or the default file can be generated by simply invoking
the routine with no parameters specified.
}
\usage{
   createparfile(filename='Lambdar_default.par',
                 RootDirectory = "./", WorkingDirectory = "./",
                 OutputDirectory = "./", Catalogue = "catalogue.csv",
                 DataMap = "image.fits", BeamArea_SqAS = 0, PSFConvolve
                 = 0, PSFWeighted = 1, PSFMap = "NONE", Gauss_FWHM_AS =
                 0, SimGauss_AS=0, RemoveContam = 0, CheckContam = 0, nNearestCheck =
                 10, NoContamImageFile = "NoContamResidualImage.fits",
                 CatIDColumnLabel = "CATAID", RAColumnLabel =
                 "RAdeg", DecColumnLabel = "DECdeg",
                 ThetaColumnLabel = "rotN2E", SemiMajColumnLabel =
                 "radmajasec", SemiMinColumnLabel = "radminasec",
                 ContamColumnLabel = "CONTAM", FluxWgtColumnLabel =
                 "FLUXWEIGHT", ErrorMap = "NONE", MaskMap = "NONE",
                 WeightMap = "NONE", WeightMapZP = 0, Saturation = Inf,
                 SaturationLabel = "SATUR", GainLabel = "GAIN",
                 DataExtn = 0, ErrorExtn = 0, MaskExtn = 0, WeightExtn
                 = 0, PointSources = 0, EFactor = 0, FluxCorr = 1,
                 CropImage = 0, CropFitsName = "croppedimage",
                 CropImRA0 = -999, CropImDec0 = -999, CropImRad = 0.5,
                 Confusion_units = 0, nProcessors = 1, AngularOffset = 0,
                 MapUnitsPerBeam = 0, SmoothAper = 0, ResamplingRes = 3,
                 ResamplingIters = 2, PSFConfidence = 1, ApStampWidth =
                 1.05, SourceMaskOnly = 0, WriteSourceMask = 0,
                 WriteAAMask = 0, AllApersFile =
                 "AllApertures_Mask.fits", WriteFAMask = 0,
                 ConvApersFile = "AllConvolvedApertures_Mask.fits",
                 WriteDFAMask = 0, DeblConvApersFile =
                 "AllDeblConvolvedApertures_Mask.fits", WriteResidMap =
                 0, ResidImageFile = "ResidualImage.fits", WriteTable =
                 1, TableName = "LAMBDAR_Results", ShowTime = 0,
                 Interactive = 0, UseMaskLim = 0.2, Diagnostic = 0,
                 Verbose = 0, PlotSample = 0, PlotAll = 0, Magnitudes =
                 1, MagZPLabel = "MagZP", ABVegaFlux = 1, MagZeroPoint
                 = 8.9, BlankCor = 0, nBlanks = 10, RanCor = 0, nRandoms
                 = 10, DoSkyEst = 0, GetSkyRMS = 0, SkyEstIters = 5,
                 SkyEstProbCut = 3, SkyCorrelNoise = 1, SkyDefault =
                 "median", SourceMaskFile = "SourceMask.fits",
                 TransmissionMap = 0, GetDeblFrac = 0,
                 SourceMaskConfLim = 0.95, MinApRad = 0, MemorySafe =
                 0, ApertureConfLimit = 0.9, IterateFluxes = 0,
                 nIterations = 2, FluxWgtType = "scale",
                 UsePixelFluxWgts = 0, LogFile = "LAMBDAR_Log.txt")
}
%- maybe also `usage' for other objects documented here.
\arguments{
 \item{filename}{ string; Filename to use for outputting Lambdar Parameter File}
 \item{RootDirectory}{ string; Path to Root Directory for program execution. Default is `./'; i.e. the current R working directory}
 \item{WorkingDirectory}{ string; Path of Working Directory, relative to the Root Directory, for program execution.}
 \item{Catalogue}{ string; Catalogue Filename and Path from Root. Accepted formats are `.csv', `.fits', `.Rdata' }
 \item{PSFMap}{ string; PSF Filename and Path from Root (use `NONE' if not wanted) }
 \item{DataMap}{ string; Image Filename and Path from Root }
 \item{ErrorMap}{ string; Error Map Filename and Path from Root (use `NONE' if not wanted) }
 \item{MaskMap}{ string; Mask Filename and Path from Root (use `NONE' if not wanted) }
 \item{WeightMap}{ string; Weightmap Filename and Path from Root (used if generating a mask from weightmap - use `NONE' if not wanted) }
 \item{CatIDColumnLabel}{ string; Label for `Catalogue ID' Column in input catalogue }
 \item{RAColumnLabel}{ string; Label for `Right Ascension' Column in input catalogue }
 \item{DecColumnLabel}{ string; Label for `Declination' Column in input catalogue }
 \item{ThetaColumnLabel}{ string; Label for `Aperture Angle' Column in input catalogue (column may not be needed)}
 \item{SemiMinColumnLabel}{ string; Label for `SemiMinor Axis' Column in input catalogue (column may not be needed)}
 \item{SemiMajColumnLabel}{ string; Label for `SemiMajor Axis' Column in input catalogue (column may not be needed)}
 \item{ContamColumnLabel}{ string; Label for `Contaminant Flag' Column in input catalogue (column may not be needed)}
 \item{FluxWgtColumnLabel}{ string; Label for `Flux Weights' Column in input catalogue (column may not be needed)}
 \item{FluxWgtType}{ string; What form do the fluxweights take? Magnitudes `mag', Fluxes `flux', or scales `scale'. }
 \item{MapUnitsPerBeam}{ boolean 0/1; Is the Image Flux measured in Unitsperbeam (as opposed to UnitsPerPixel)? [1/0] }
 \item{Confusion_units}{ float; Confusion Correction in Units }
 \item{Saturation}{ float; image saturation in ADU }
 \item{SaturationLabel}{ string; saturation fits keyword }
 \item{GainLabel}{ string; gain fits keyword. Used to determine gain if no weightmap or errormap is provided  }
 \item{Gauss_FWHM_AS}{ float; Gaussian FWHM of Seeing in ArcSec (used if generating a PSF)}
 \item{SimGauss_AS}{ float; FWHM of Gaussian used when performing Gaussianisation of the input image. Used when generating a Simulated Image,
 to accurately reproduce the background correlations. }
 \item{BeamArea_SqAS}{ float; Beam area in Square Arcsec (used if MapUnitsPerBeam=1)}
 \item{FluxCorr}{ float; Flux Correction Factor (multiplicative) }
 \item{EFactor}{ float; Error Map Scale Factor (multiplicative) }
 \item{DataExtn}{ integer; Fits header extension of Data in Image File }
 \item{ErrorExtn}{ integer; Fits header extension of Data in Errormap File (used if reading errormap from file)}
 \item{MaskExtn}{ integer; Fits header extension of Data in Maskmap File (used if reading mask from file)}
 \item{WeightExtn}{ integer; Fits header extension of Data in Weightmap File (used if generating mask from weightmap)}
 \item{WeightMapZP}{ float; Zero Point of the Weight Map. i.e. the value that corresponds to `No Data' (used if generating mask) }
 \item{CropImage}{ boolean 0/1; Do we want to crop the input image to a particular subsection? [1/0] }
 \item{CropImRA0}{ float; RA that crop will centre around (deg; -999 will focus crop around the centre of the input image) }
 \item{CropImDec0}{ float; Dec that crop will centre around (deg; -999 will focus crop around the centre of the input image) }
 \item{CropImRad}{ float; Radius of crop (deg; NA will default to 1deg radius) }
 \item{CropFitsName}{ string; Name of the output cropped images (without `.fits' extension) }
 \item{RemoveContam}{ boolean 0/1; Remove Contaminants (Catalogue must contain contaminant flag column) from the image? [1/0] }
 \item{CheckContam}{ boolean 0/1; Check for relevant Contaminants? Contaminants deemed not needed (not overlapping science targets) will not be measured. [1/0] }
 \item{nNearestCheck}{ integer; number of nearest neighbours to check when discarding contaminant }
 \item{MinApRad}{ flaot; Minimum Aperture Radius }
 \item{PSFConvolve}{ boolean 0/1; Convolve the Aperture with a PSF? [1/0] }
 \item{PSFWeighted}{ boolean 0/1; Use PSF Weighted Apertures [1] or convert all apertures to boxcar prior to measurement [0]? }
 \item{ApertureConfLimit}{ float; when not using PSF Weighted Apertures, this is the integral point at which the boxcar boundary occurs }
 \item{AngularOffset}{ boolean 0/1; 0 if the catalogue is in N90E0 angular coords, 1 if it is in N0E90 coords [0/1]}
 \item{PointSources}{ boolean 0/1; Force point sources to be used? [1/0] }
 \item{SmoothAper}{ boolean 0/1; Smooth Apertures by resampling? (Improves aperture surface integral accuracy) [1/0] }
 \item{ResamplingRes}{ integer; Rate of Upscaling in Resolution during Aperture Resampling (final res is ResamplingRes^[nIters] higher) }
 \item{ResamplingIters}{ integer; Number of iterations to perform in resampling }
 \item{UseMaskLim}{ float in \{0,1\}; fraction of aperture stamp that must be within the mask. If less, aperture is discarded. }
 \item{Verbose}{ boolean 0/1; Verbose Output? [1/0] }
 \item{ShowTime}{ boolean 0/1; Display execution times & total elapsed time during run? [1/0] }
 \item{PlotSample}{ boolean 0/1; Plot a random sample of the Apertures to file, during various procedures, for Inspection? [1/0] }
 \item{PlotAll}{ boolean 0/1; Plot all of the Apertures to file, during various procedures, for Inspection? [1/0] }
 \item{Diagnostic}{ boolean 0/1; Diagnostic Output of Variable Values During Computation - Helpful in Understanding Code & debugging [1/0] }
 \item{Interactive}{ boolean 0/1; Interactive Mode - Allows user access to entire parameter-space after final calculations [1/0] }
 \item{Magnitudes}{ boolean 0/1; Output Flux Magnitudes? [1/0] }
 \item{ABVegaFlux}{ float; Flux of Vega in this band (used for calculation of AB Magnitudes) }
 \item{MagZeroPoint}{ float; Magnitude of the Image Zero Point. If -999, ZP will be read from FITS header }
 \item{MagZPLabel}{ string; Label used for the Zero Point Keyword in the FITS header, if reading from FITS }
 \item{DoSkyEst}{ boolean 0/1; Perform a subtraction of the local background for each aperture? [1/0] }
 \item{SkyEstIters}{ integer; Number of iterative clips performed during sky estimation }
 \item{SkyEstProbCut}{ float; Probability limit (specified as a sigma value) used in iterative clipping of sky-pixels during sky estimation }
 \item{SkyDefault}{ character/float; The default value to use if sky estimation fails; can be a value (numeric), `mean', or `median' }
 \item{SkyCorrelNoise}{ float; Multiplicative factor describing the amount of correlated noise in the datamap }
 \item{GetSkyRMS}{ boolean 0/1; Perform an estimation of the local sky background, without doing subtraction }
 \item{RanCor}{ boolean 0/1; Perform a randoms measurement for each aperture? [1/0] }
 \item{nRandoms}{ integer; number of randoms to measure, per aperture. }
 \item{BlankCor}{ boolean 0/1; Perform a randoms measurement for each aperture? [1/0] }
 \item{nBlanks}{ integer; number of Blanks to attempt to measure, per aperture. Actual number successfully measured is output by blanks function}
 \item{UsePixelFluxWgts}{ boolean 0/1; Do we want to weight initialise deblends by object central pixel fluxes? [0/1] }
 \item{IterateFluxes}{ boolean 0/1; Perform an iteration of the flux measurements, to create improve photometric accuracy (generally, not always) }
 \item{nIterations}{ integer; Number of iterations we would like to perform when refining fluxes }
 \item{GetDeblFrac}{ boolean 0/1; Do we just want to return the deblend solutions? No flux measurement will be output [0/1] }
 \item{OutputDirectory}{ string; Output directory Name and Path }
 \item{LogFile}{ string; Filename for Log }
 \item{SourceMaskOnly}{ boolean 0/1; Only Output Source Mask image to file, and stop (without calculating fluxes)? (Overrules all other Flags) [1/0] }
 \item{TransmissionMap}{ boolean 0/1; Do we want to output a transmission map rather than a sourcemask? [1/0] }
 \item{WriteTable}{ boolean 0/1; Output Results Table? [1/0]}
 \item{TableName}{ string; Filename for Results Table }
 \item{WriteAAMask}{ boolean 0/1; Write the All Apertures Mask image to file? [1/0] }
 \item{AllApersFile}{ string; Filename for All Apertures Mask }
 \item{WriteSourceMask}{ boolean 0/1; Write the Source Mask image to file? [1/0] }
 \item{SourceMaskConfLim}{ float in \{0,1\}; Confidence Limit used when making sourcemask of convolved apertures }
 \item{SourceMaskFile}{ string; Filename for Source Mask }
 \item{WriteDFAMask}{ boolean 0/1; Write the Convolved Apertures Mask image to file? [1/0] }
 \item{DeblConvApersFile}{ string; Filename for All Convolved Apertures Mask }
 \item{WriteFAMask}{ boolean 0/1; Write the Convolved Apertures Mask image to file? [1/0] }
 \item{ConvApersFile}{ string; Filename for All Convolved Apertures Mask }
 \item{WriteResidMap}{ boolean 0/1; Write the Residual Image Map to file? [1/0] }
 \item{ResidImageFile}{ string; Filename for Residual Image }
 \item{NoContamImageFile}{ string; Filename for Contaminant Subtracted Residual Image }
 \item{nProcessors}{ integer; Number of Processors Available for Use in Computations }
 \item{MemorySafe}{ boolean 0/1; do we want to check that memory usage will be less than system limits? [0/1] }
 \item{ApStampWidth}{ float in \{1,Inf\}; Width of the Aperture stamps in multiples of aperture major axis; Can be changed *with caution* if memory issues arise. }
 \item{PSFConfidence}{ float in \{0,1\}; Confidence level to which the PSF is generated, or to which aperture stamps are padded if reading PSF from File}
}
\details{
This function reads the various inputs from the LAMBDAR parameter file. Generally, if a particular parameter isn't found during parameter file read
then the program will warn the user in the log file, and will continue on using the default values for that parameter. In this way a user that wishes to
use all default values, and specify only one or two parameters explicitly, can provide a simply file containing only those parameters that they wish to
modify.
Exceptions to this rule are those parameters that are considered vital to the correct running of the program. These are:
  > The Root Path (can be simply set to "./")
  > The Working Directory Path (can be simply set to "./")
  > The Input Image File
  > The Input Catalogue

Details of what is expected in the input catalogue are given in the \link{opencatalogue} manual page
}
\value{
NULL; Procedure outputs the parameter file to `LAMBDAR_default.par'
}
\references{
Wright et al. 2015 (in prep)
}
\author{
Angus H Wright{ ICRAR}
 {\email{angus.wright@icrar.org}}
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\link{MeasureFluxes}
}
\examples{
#To output the default parameter file to 'Lambdar_default.par', simply call:
createparfile()

#To specify the output file as 'Lambdar_Custom.par':
createparfile(file="Lambdar_Custom.par")

#Finally, all parameters can be set as they are output:
createparfile(file="Lambdar_Custom.par", Catalogue="MyCat.csv", DataMap="SampleImage.fits",nProcessors=10)

}
% Add one or more standard keywords, see file `KEYWORDS' in the
% R documentation directory.
\keyword{ Parameters }
\keyword{ Output }
