% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/Filter_.R
\name{Filter_}
\alias{Filter_}
\title{Filter function for lambda expression}
\usage{
Filter_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}
}
\description{
Filter function for lambda expression
}

