% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Find_}
\alias{Find_}
\title{Find function for lambda expression}
\usage{
Find_(data, ..., right = FALSE, nomatch = NULL)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}
}
\value{
an index.
}
\description{
Find function for lambda expression
}

