% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-tag.R
\name{get_handler_from_block}
\alias{get_handler_from_block}
\title{Extract handler functions from a \code{roxy_block}}
\usage{
get_handler_from_block(block)
}
\arguments{
\item{block}{A \code{roxy_block}}
}
\value{
A named character vector containing \code{file} and \code{function_name} items.
}
\description{
Extract handler functions from a \code{roxy_block}
}
\keyword{internal}
