% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{What operating system are we on?}
\usage{
get_os()
}
\value{
One of \code{"osx"}, \code{"linux"} or \code{"windows"}.
}
\description{
Needed to warn Windows users
}
\keyword{internal}
