% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{init}
\alias{init}
\title{Use lambdar with your project}
\usage{
init()
}
\description{
Use lambdar with your project
}
\section{Description}{

Call this function to create the necessary files lambdar needs to work. Specifically:
\itemize{
\item A \verb{_lambdar.yml} config file
\item A \verb{lambdar/} directory to house a copy of the R runtime we will load into your container
}
}

