% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_aws_get_account_id}
\alias{lam_aws_get_account_id}
\title{Attempt to get our AWS account ID}
\usage{
lam_aws_get_account_id()
}
\value{
A character string
}
\description{
Attempt to get our AWS account ID
}
\keyword{internal}
