% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{lam_build_dockerfile}
\alias{lam_build_dockerfile}
\title{Build a Dockerfile}
\usage{
lam_build_dockerfile(cfg)
}
\arguments{
\item{cfg}{A [lambdar_config] object.}
}
\value{
Nothing - this function is called for its side effect, which is to write a `Dockerfile`
  to disk.
}
\description{
Build a Dockerfile
}
\keyword{internal}
