% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{lam_docker_image_exists}
\alias{lam_docker_image_exists}
\title{Check whether an image with the specified name exists on the system}
\usage{
lam_docker_image_exists(image_name, tag = NULL)
}
\arguments{
\item{image_name}{Name of the image. Can optionally include a tag, like \code{"my-image:latest"}.}

\item{tag}{Optional image tag, i.e. \code{{image_name}:{tag}}. If a tag is provided in \code{image_name}
it takes precedence and this argument is ignored.}
}
\value{
Boolean
}
\description{
Check whether an image with the specified name exists on the system
}
\keyword{internal}
