% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{lam_ecr_create_repo_if_not_exists}
\alias{lam_ecr_create_repo_if_not_exists}
\title{Create an ECR repository if it doesn't already exist}
\usage{
lam_ecr_create_repo_if_not_exists(aws_ecr_repository_name)
}
\arguments{
\item{aws_ecr_repository_name}{Chosen name for your AWS ECR repository. It is recommended that
this be the same as your app name.}
}
\description{
Create an ECR repository if it doesn't already exist
}
\keyword{internal}
