% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{lam_ecr_image_tag_name}
\alias{lam_ecr_image_tag_name}
\title{Create the AWS ECR tag for your image}
\usage{
lam_ecr_image_tag_name(
  aws_account_id,
  aws_region,
  aws_ecr_repository_name,
  tag = "latest"
)
}
\arguments{
\item{aws_account_id}{Your 12-digit AWS account id}

\item{aws_region}{Your AWS region, e.g. \code{"ap-southeast-2"}}

\item{aws_ecr_repository_name}{Chosen name for your AWS ECR repository. It is recommended that
this be the same as your app name.}

\item{tag}{Optional image tag. If omited, "latest" will be used.}
}
\value{
A character vector
}
\description{
Create the AWS ECR tag for your image
}
\keyword{internal}
