% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_ecr_repo_url}
\alias{lam_ecr_repo_url}
\title{Generate an ECR repo URL}
\usage{
lam_ecr_repo_url(aws_account_id, aws_region, aws_ecr_repository_name)
}
\arguments{
\item{aws_account_id}{Your 12-digit AWS account id}

\item{aws_region}{Your AWS region, e.g. \code{"ap-southeast-2"}}

\item{aws_ecr_repository_name}{Chosen name for your AWS ECR repository. It is recommended that
this be the same as your app name.}
}
\description{
Generate an ECR repo URL
}
\keyword{internal}
