% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_ecr_tag_image_for_upload}
\alias{lam_ecr_tag_image_for_upload}
\title{Tag an image in preparation for upload to AWS Lambda}
\usage{
lam_ecr_tag_image_for_upload(tag = "latest")
}
\arguments{
\item{tag}{Optional image tag. If omited, "latest" will be used.}
}
\description{
See \href{https://aws.amazon.com/blogs/aws/new-for-aws-lambda-container-image-support/}{this blog post}
for details.
}
\keyword{internal}
