% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{lam_ecr_upload_image}
\alias{lam_ecr_upload_image}
\title{Upload an image to ECR}
\usage{
lam_ecr_upload_image(aws_account_id, aws_region, aws_ecr_repository_name)
}
\arguments{
\item{aws_account_id}{Your 12-digit AWS account id}

\item{aws_region}{Your AWS region, e.g. \code{"ap-southeast-2"}}

\item{aws_ecr_repository_name}{Chosen name for your AWS ECR repository. It is recommended that
this be the same as your app name.}
}
\description{
Upload an image to ECR
}
\keyword{internal}
