% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_find_sourced_files}
\alias{lam_find_sourced_files}
\title{Find all the files \code{source}d from another}
\usage{
lam_find_sourced_files(file)
}
\arguments{
\item{file}{A file path}
}
\value{
A character vector of files \code{source}d in \code{file}.
}
\description{
Find all the files \code{source}d from another
}
\keyword{internal}
