% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_function_exists_in_file}
\alias{lam_function_exists_in_file}
\title{Check that a file contains a function}
\usage{
lam_function_exists_in_file(file, fun)
}
\arguments{
\item{file}{A file path}

\item{fun}{String. A function that we expect to be defined in \code{file}}
}
\value{
Boolean
}
\description{
Check that a file contains a function
}
\keyword{internal}
