% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_get_file_dependencies}
\alias{lam_get_file_dependencies}
\title{Detect R package dependencies in code}
\usage{
lam_get_file_dependencies(file)
}
\arguments{
\item{file}{Path to a file or files to analyse}
}
\value{
A character vector of R package dependencies needed by \code{file}
}
\description{
Detect R package dependencies in code
}
\keyword{internal}
