% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_handler_filenames}
\alias{lam_handler_filenames}
\title{Get a list of all \code{.R} files containing \verb{@lambda} handlers}
\usage{
lam_handler_filenames()
}
\value{
Character vector of file paths
}
\description{
Get a list of all \code{.R} files containing \verb{@lambda} handlers
}
\keyword{internal}
