% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{lam_has_aws_cli}
\alias{lam_has_aws_cli}
\title{Do we have the AWS cli installed?}
\usage{
lam_has_aws_cli()
}
\value{
Boolean
}
\description{
Examines the output of \code{\link[=Sys.which]{Sys.which()}} to see if we can locate an installation.
}
\keyword{internal}
