% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{lam_has_docker}
\alias{lam_has_docker}
\title{Do we have docker installed?}
\usage{
lam_has_docker()
}
\value{
Boolean
}
\description{
Examines the output of \code{\link[=Sys.which]{Sys.which()}} to see if we can locate an installation.
}
\keyword{internal}
