% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-tag.R
\name{lam_parse_file_handlers}
\alias{lam_parse_file_handlers}
\title{Parse files containing \verb{@lambda} tags into AWS Lambda handler specifiers}
\usage{
lam_parse_file_handlers(file)
}
\arguments{
\item{file}{A file path}
}
\value{
Character vector containing \code{"file.function_name"} for any \verb{@lambda}-tagged functions in
\code{file}.
}
\description{
AWS Lambda handlers are specified as \code{file.function_name}.
}
\keyword{internal}
