% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_proj_path}
\alias{lam_proj_path}
\alias{lam_config_file}
\alias{lam_dockerfile_path}
\title{Path utilities}
\usage{
lam_proj_path(...)

lam_config_file()

lam_dockerfile_path()
}
\arguments{
\item{...}{Path elements}
}
\description{
Pinched from `usethis`.
}
\section{Functions}{
\itemize{
\item \code{lam_config_file}: Path to lambdar config file

\item \code{lam_dockerfile_path}: Path to Dockerfile
}}

\keyword{internal}
