% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_r_version}
\alias{lam_r_version}
\title{Get the current R version}
\usage{
lam_r_version()
}
\value{
Current R version as a string, e.g. "4.0.1"
}
\description{
Get the current R version
}
\keyword{internal}
