% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_tag.R
\name{lam_roxy_block_has_lambda_tag}
\alias{lam_roxy_block_has_lambda_tag}
\title{Does a parsed roxygen block contain a \verb{@lambda} tag?}
\usage{
lam_roxy_block_has_lambda_tag(block)
}
\arguments{
\item{block}{A roxy_block from \code{\link[roxygen2:parse_package]{roxygen2::parse_file()}}}
}
\value{
\code{TRUE} if the block contains a \verb{@lambda} tag, \code{FALSE} otherwise.
}
\description{
Does a parsed roxygen block contain a \verb{@lambda} tag?
}
\keyword{but}
\keyword{for}
\keyword{how}
\keyword{internal}
\keyword{it}
\keyword{now.}
\keyword{properly,}
\keyword{this}
\keyword{to}
\keyword{understand}
\keyword{use}
\keyword{works}
