% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-utils.R
\name{lam_run_system_command}
\alias{lam_run_system_command}
\title{Invoke a system command and check to see if it was successful}
\usage{
lam_run_system_command(cmd, capture_output = FALSE, quiet = lam_is_quiet())
}
\arguments{
\item{cmd}{String, a system command}

\item{capture_output}{Do we want the actual output of the command returned? Useful when trying to
generate text. This is passed directly to tne \code{intern} parameter of \code{\link[=system]{system()}}.}

\item{quiet}{Set to \code{FALSE} to reduce verbosity of lambdar messages. Note that system commands
may still write to the terminal.}
}
\value{
If \code{capture_output} is \code{FALSE}, invisibly returns the exit code of the command, as long
as that is zero. Any other exit code throws an error. If \code{capture_output} is \code{TRUE}, returns
whatever the text output of the command is.
}
\description{
This is a wrapper around \code{\link[=system]{system()}} with some extra bells and whistles.
}
\keyword{internal}
