% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_scan_project}
\alias{lam_scan_project}
\title{Scan the project folder for handlers and package dependencies}
\usage{
lam_scan_project()
}
\value{
A list containing three named items, \code{lambda_handlers}, \code{include_files} and \code{r_packages},
each of which is a character vector, possible of length zero.
}
\description{
Scan the project folder for handlers and package dependencies
}
\keyword{internal}
