% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{Lambda function}
\usage{
lambda(..., envir = parent.frame())
}
\arguments{
\item{...}{lambda expression.}

\item{envir}{environment.}
}
\description{
Lambda function
}

