% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar_config.R
\name{lambdar_config}
\alias{lambdar_config}
\alias{new_lambdar_config}
\alias{is_lambdar_config}
\title{Lambdar configuration object}
\usage{
lambdar_config(config_file = NULL)

new_lambdar_config(cfg)

is_lambdar_config(x)
}
\arguments{
\item{config_file}{Path to config file. If not provided, looks for \verb{_lambdar.yml} in the project
root durectory.}
}
\value{
A list.
}
\description{
Contains info about the container
}
\section{Functions}{
\itemize{
\item \code{new_lambdar_config}: Create a new \code{lambdar_config} object. For internal use only

\item \code{is_lambdar_config}: Test if an object is a \code{lambdar_config} object.
}}

\keyword{internal}
