% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar-config.R
\name{lambdar_config_from_file}
\alias{lambdar_config_from_file}
\alias{lambdar_config}
\alias{new_lambdar_config}
\alias{is_lambdar_config}
\title{Lambdar configuration object}
\usage{
lambdar_config_from_file(config_file = NULL)

new_lambdar_config(config = NULL)

is_lambdar_config(x)
}
\arguments{
\item{config_file}{Path to config file. If not provided, looks for \verb{_lambdar.yml} in the project
root durectory.}

\item{config}{A named list containing params. If \code{config} or any elements are missing, defaults
will be used.}
}
\value{
A list.
}
\description{
Lambdar stores information about your app in a configuration object of class \code{lambdar_config}.
This is just a named list.
}
\section{Functions}{
\itemize{
\item \code{new_lambdar_config}: Create a new \code{lambdar_config} object. For internal use only.

\item \code{is_lambdar_config}: Test if an object is a \code{lambdar_config} object.
}}

\keyword{internal}
