% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map2_.R
\name{Map2_}
\alias{Map2_}
\title{Map function for lambda expression}
\usage{
Map2_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression.}
}
\value{
a list.
}
\description{
Map function for lambda expression
}

