% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Map2_.R
\name{Map2v_}
\alias{Map2v_}
\title{Map function for lambda expression}
\usage{
Map2v_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression.}
}
\value{
a vector.
}
\description{
Map function for lambda expression
}

