% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/Map_.R
\name{Map_}
\alias{Map_}
\title{Map function for lambda expression}
\usage{
Map_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression.}
}
\description{
Map function for lambda expression
}

