% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Mapu_}
\alias{Mapu_}
\title{Map function for lambda expression with unlisted output}
\usage{
Mapu_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression.}
}
\value{
a vector.
}
\description{
Map function for lambda expression with unlisted output
}

