% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_.R
\name{Mapv_}
\alias{Mapu_}
\alias{Mapv_}
\title{Map function for lambda expression with unlisted output}
\usage{
Mapv_(data, ...)

Mapu_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression.}
}
\value{
a vector.
}
\description{
Map function for lambda expression with unlisted output
}

