% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Negate_}
\alias{Negate_}
\title{Negate function for lambda expression}
\usage{
Negate_(...)
}
\arguments{
\item{...}{lambda expression(unary).}

\item{data}{a vector.}
}
\value{
a function.
}
\description{
Negate function for lambda expression
}

