% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Position_}
\alias{Position_}
\title{Position function for lambda expression}
\usage{
Position_(data, ..., right = FALSE, nomatch = NA_integer_)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}
}
\value{
an index.
}
\description{
Position function for lambda expression
}

