% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar-config.R
\name{read_config}
\alias{read_config}
\title{Read a YAML file and return a config object}
\usage{
read_config()
}
\value{
A \link{lambdar_config} object.
}
\description{
Read a YAML file and return a config object
}
