% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-tag.R
\name{roclet_lambda}
\alias{roclet_lambda}
\title{Roclet lambda class}
\usage{
roclet_lambda()
}
\description{
Create an object of class \code{roclet_lambda} to ensure the correct S3 methods are called.
}
\keyword{internal}
