% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_tag.R
\name{roclet_process.roclet_lambda}
\alias{roclet_process.roclet_lambda}
\title{Extracts functions to be lambda-d from the roxygen block/s}
\usage{
\method{roclet_process}{roclet_lambda}(x, blocks, env, base_path)
}
\arguments{
\item{x}{A \code{roclet_lambda} object. See \code{\link[=roclet_lambda]{roclet_lambda()}}.}

\item{blocks}{A list of \link[roxygen2:roxy_block]{roxygen2::roxy_block}s. Each block corresponds to the captured into of
one function.}

\item{env}{An environment - not sure what this does.}

\item{base_path}{Not sure about this either.}
}
\description{
Extracts functions to be lambda-d from the roxygen block/s
}
\keyword{internal}
