% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{use_lambdar}
\alias{use_lambdar}
\title{Use lambdar with your project}
\usage{
use_lambdar()
}
\description{
Use lambdar with your project
}
\section{Description}{

Call this function to create the necessary files lambdar needs to work. Specifically:

* A `_lamndar.yml` config file
* A `lambdar/` directory to house a copy of the R runtime we will load into your container.

The function will try and pre-populate your config file with as much information as possible, but
you are free to amend as much or as little as you like.
}

