% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R, R/lambdar.R
\name{use_lambdar_yaml}
\alias{use_lambdar_yaml}
\title{Add `_lamdar.yml` to the project root.}
\usage{
use_lambdar_yaml(save_as = "_lambdar.yml")

use_lambdar_yaml(save_as = "_lambdar.yml")
}
\description{
Add `_lamdar.yml` to the project root.

Add `_lamdar.yml` to the project root.
}
\keyword{internal}
