% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{using_lambdar}
\alias{using_lambdar}
\alias{runtime_exists}
\alias{config_exists}
\alias{dockerfile_exists}
\title{Have we got everything we need?}
\usage{
using_lambdar()

runtime_exists()

config_exists()

dockerfile_exists()
}
\value{
Boolean
}
\description{
Have we got everything we need?
}
\section{Functions}{
\itemize{
\item \code{runtime_exists}: Does the runtime file exist?

\item \code{config_exists}: Does the config file exist?

\item \code{dockerfile_exists}: Does the Dockerfile exist?
}}

\keyword{internal}
