% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-utils.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Take a config object, format its contents and write the Dockerfile}
\usage{
write_dockerfile(cfg)
}
\arguments{
\item{cfg}{A \link{lambdar_config} object.}
}
\description{
Formatting involves correctly quoting R vectors for inclusion in the config file.
}
\keyword{internal}
