% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{build_dockerfile}
\alias{build_dockerfile}
\title{Create a Dockerfile}
\usage{
build_dockerfile(cfg = NULL, quiet = FALSE)
}
\arguments{
\item{cfg}{A [lambdar_config()] object (optional).}

\item{quiet}{If `FALSE`, will tell you the command to build your container.}
}
\description{
Writes a Dockerfile to your project root directory based on a configuration object. If no
config object is provided it will try and read `_lambdar.yml` from your project's root directory.
}
