% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{build_dockerfile}
\alias{build_dockerfile}
\title{Create a Dockerfile}
\usage{
build_dockerfile()
}
\value{
A \link{lambdar_config} object, invisibly. This is for internal use only and may be removed in
future.
}
\description{
Writes a Dockerfile to your project root directory based on a configuration file If no config
file path is provided it will try and read \verb{_lambdar.yml} from your project's root directory.
}
