% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{build_yaml}
\alias{build_yaml}
\title{Build a \verb{_lambdar.yml} file}
\usage{
build_yaml()
}
\description{
Scans your project folder for \code{.R} files, identified those containing \verb{@lambda} tags, and writes
a \verb{_lambdar.yml} config file containing relevant information.
}
