% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Filter_}
\alias{Filter_}
\title{Filter function for lambda expression}
\usage{
Filter_(data, ...)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}
}
\value{
a vector.
}
\description{
Filter function for lambda expression
}

