% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_.R
\name{Find_}
\alias{Find_}
\title{Find function for lambda expression}
\usage{
Find_(data, ..., right = FALSE, nomatch = NULL)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}

\item{right}{a logical indicating whether to proceed from left to right (default) or from right to left.}

\item{nomatch}{the value to be returned in the case when "no match" (no element satisfying the predicate) is found.}
}
\value{
an obejct.
}
\description{
Find function for lambda expression
}

