% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar.R
\name{lam_build_env_list}
\alias{lam_build_env_list}
\title{Create a space-separated list of environment variables}
\usage{
lam_build_env_list(env = list())
}
\arguments{
\item{env}{A named list}
}
\description{
Create a space-separated list of environment variables
}
