% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_get_project_dependencies}
\alias{lam_get_project_dependencies}
\title{Detect R package dependencies in code}
\usage{
lam_get_project_dependencies()
}
\value{
A character vector of R package dependencies needed by \code{file}
}
\description{
Uses \code{\link[renv:dependencies]{renv::dependencies()}} to work out what packages we depend on. Ignores \code{rmarkdown} where
that is triggered by the presence of markdown files, and also excludes packages that are already
in the container because the runtime needs them (\code{httr}, \code{jsonlite} and \code{logger}).
}
\keyword{internal}
