% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lam_proj_path}
\alias{lam_proj_path}
\alias{lam_dir_path}
\alias{lam_config_path}
\alias{lam_dockerfile_path}
\alias{lam_runtime_path}
\title{Path utilities}
\usage{
lam_proj_path(...)

lam_dir_path()

lam_config_path()

lam_dockerfile_path()

lam_runtime_path()
}
\arguments{
\item{...}{Path elements}
}
\description{
Pinched from \code{usethis}.
}
\section{Functions}{
\itemize{
\item \code{lam_dir_path}: Path to the \verb{.lambdar/} directory

\item \code{lam_config_path}: Path to \verb{_lambdar.yml} config file

\item \code{lam_dockerfile_path}: Path to Dockerfile

\item \code{lam_runtime_path}: Path to runtime function
}}

\keyword{internal}
