% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{f_}
\alias{lambda}
\title{Lambda function}
\usage{
lambda(..., envir = parent.frame())

f_(..., envir = parent.frame())
}
\arguments{
\item{...}{lambda expression.}

\item{envir}{environment.}
}
\description{
Lambda function
}

