% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar-config.R
\name{lambdar_config_from_file}
\alias{lambdar_config_from_file}
\alias{lambdar_config}
\alias{new_lambdar_config}
\alias{is_lambdar_config}
\title{Create a new lambdar config object by reading the config file}
\usage{
lambdar_config_from_file()

new_lambdar_config(config = NULL)

is_lambdar_config(x)
}
\arguments{
\item{config}{A named list containing params. If \code{config} or any elements are missing, defaults
will be used.}
}
\value{
A \link{lambdar_config} object.
}
\description{
Create a new lambdar config object by reading the config file
}
\section{Functions}{
\itemize{
\item \code{new_lambdar_config}: Create a new \code{lambdar_config} object. For internal use only.

\item \code{is_lambdar_config}: Test if an object is a \code{lambdar_config} object.
}}

\keyword{internal}
