% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Negate_.R
\name{Negate_}
\alias{Negate_}
\title{Negate function for lambda expression}
\usage{
Negate_(...)
}
\arguments{
\item{...}{lambda expression.}
}
\value{
a function.
}
\description{
Negate function for lambda expression
}

