% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Position_.R
\name{Position_}
\alias{Position_}
\title{Position function for lambda expression}
\usage{
Position_(data, ..., right = FALSE, nomatch = NA_integer_)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(unary).}

\item{right}{a logical indicating whether to proceed from left to right (default) or from right to left.}

\item{nomatch}{the value to be returned in the case when "no match" (no element satisfying the predicate) is found.}
}
\value{
an index.
}
\description{
Position function for lambda expression
}

