% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_.R
\name{Reduce_}
\alias{Reduce_}
\title{Reduce function for lambda expression}
\usage{
Reduce_(data, ..., init, right = FALSE, accumulate = FALSE)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(binary).}

\item{init}{an R object of the same kind as the elements of x.}

\item{right}{a logical indicating whether to proceed from left to right (default) or from right to left.}

\item{accumulate}{a logical indicating whether the successive reduce combinations should be accumulated. By default, only the final combination is used.}
}
\value{
a value.
}
\description{
Reduce function for lambda expression
}

