% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Reduce_}
\alias{Reduce_}
\title{Reduce function for lambda expression}
\usage{
Reduce_(data, ..., init, right = FALSE, accumulate = FALSE)
}
\arguments{
\item{data}{a vector.}

\item{...}{lambda expression(binary).}
}
\description{
Reduce function for lambda expression
}

