% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relish}
\alias{relish}
\title{Tidy file paths by removing instances of \code{dir} from \code{x}.}
\usage{
relish(x, dir = getwd())
}
\arguments{
\item{x}{File path}

\item{dir}{Directory to strip from \code{x}}
}
\value{
Character vector
}
\description{
Ensures that we get a relative path for \code{x}. Pinched from \code{rlang:::relish()}.
}
\keyword{internal}
