% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_tag.R
\name{roclet_output.roclet_lambda}
\alias{roclet_output.roclet_lambda}
\title{Do stuff with parsed lambda functions}
\usage{
\method{roclet_output}{roclet_lambda}(x, results, base_path, ...)
}
\arguments{
\item{x}{A `roclet_lambda` object. See [roclet_lambda()].}

\item{results}{The output of [roclet_process.roclet_lambda].}

\item{base_path}{Unknown.}

\item{...}{Unused}
}
\description{
Given a list of functions that have been tagged with `@lambda`, do something cool with them
}
