% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdar-config.R
\name{write_config}
\alias{write_config}
\title{Take a config object, format its contents and write the config file}
\usage{
write_config(cfg)
}
\arguments{
\item{cfg}{A \link{lambdar_config} object.}
}
\description{
Formatting involves correctly quoting R vectors for inclusion in the config file.
}
\keyword{internal}
